ticket-tracker/requirements.txt
# Core dependencies
pydantic>=1.9.0,<2.0.0
psycopg2-binary>=2.9.3
typer>=0.7.0
schedule>=1.1.0

# Database
SQLAlchemy>=1.4.31
psycopg2-binary>=2.9.3
alembic>=1.7.7

# Google Sheets integration
gspread>=5.4.0
google-auth>=2.6.0
oauth2client>=4.1.3

# HTTP and APIs
aiohttp>=3.8.1
requests>=2.27.1
tenacity>=8.0.1

# Data processing
numpy>=1.22.2
pandas>=1.4.1

# Logging and formatting
rich>=12.0.0
structlog>=21.5.0

# Concurrency
asyncio>=3.4.3

# Celery
celery>=5.2.7
flower>=1.2.0
redis>=4.3.4